Move the prometheus monitoring for fill feed to the calling script
