+++ client/ts/src/fillFeed.ts
@@ -6,8 +6,6 @@ import { PROGRAM_ID } from './manifest';
 import { convertU128 } from './utils/numbers';
 import { genAccDiscriminator } from './utils/discriminator';
 import * as promClient from 'prom-client';
-import express from 'express';
-import promBundle from 'express-prom-bundle';
 import { FillLogResult } from './types';
 
 // For live monitoring of the fill feed. For a more complete look at fill
@@ -180,40 +178,6 @@ export class FillFeed {
       });
     }
   }
-}
-
-/**
- * Run a fill feed as a websocket server that clients can connect to and get
- * notifications of fills for all manifest markets.
- */
-export async function runFillFeed() {
-  const connection: Connection = new Connection(
-    process.env.RPC_URL || 'http://127.0.0.1:8899',
-    'confirmed',
-  );
-
-  promClient.collectDefaultMetrics({
-    labels: {
-      app: 'fillFeed',
-    },
-  });
-
-  const register = new promClient.Registry();
-  register.setDefaultLabels({
-    app: 'fillFeed',
-  });
-  const metricsApp = express();
-  metricsApp.listen(8080);
-
-  const promMetrics = promBundle({
-    includeMethod: true,
-    metricsApp,
-    autoregister: false,
-  });
-  metricsApp.use(promMetrics);
-
-  const fillFeed: FillFeed = new FillFeed(connection);
-  await fillFeed.parseLogs();
 }
 
 const fillDiscriminant = genAccDiscriminator('manifest::logs::FillLog');
