+++ debug-ui/package.json
@@ -26,8 +26,11 @@
     "@solana/wallet-adapter-react-ui": "^0.9.35",
     "@solana/wallet-adapter-wallets": "^0.19.32",
     "@solana/web3.js": "^1.95.0",
+    "express": "^4.21.0",
+    "express-prom-bundle": "^7.0.0",
     "lightweight-charts": "^4.2.0",
     "next": "14.2.5",
+    "prom-client": "^15.1.3",
     "react": "^18.2.0",
     "react-dom": "^18",
     "react-toastify": "^10.0.5"
